requirements.txt
paho-mqtt >= 1.5.0
hecuba >= 0.1.3.post1
pydantic >=1.4